Chroma Service Refactor
