+++ apps/chroma_db/Readme.md
@@ -4,21 +4,15 @@ cd apps/chroma_db
 ```shell
 pip install -r requiremets.txt
 ```
-# Set up Posgres
+# Set up Posgres, Chroma
 ```shell
-docker run --name chroma_postgres   -e POSTGRES_USER=k1   -e POSTGRES_PASSWORD=k1   -e POSTGRES_DB=k1   -p 5435:5432   -d postgres:15
+docker compose up -d
 ```
 ### Note: docker ps -a to check avoid same port. if re map, adjust in config postgres
 ## create table: 
 ```shell
-python ids/postges/postgres.py
+python postges/postgres.py
 ```
-# Set up Chroma
-## 1. create container (server)
-```shell
-docker run -d -p 8000:8000 ghcr.io/chroma-core/chroma:latest
-```
-###    Note: same setup postgres
 # API
 ```shell
 python test.py
@@ -33,4 +27,4 @@ python get_ebd.py
 ```shell
 docker exec -it <id container postgre> bin/bash
 ```
-![alt text](image.png)
+![alt text](demo.png)
