+++ apps/chroma_db/configs/config.yml
@@ -1,30 +1,22 @@
-## để test nhanh
-data:
-  raw_documents_path: "./data/raw_documents"
-
 # chunk
 text_splitter:
   chunk_size: 500
   chunk_overlap: 50
 
-# Cấu hình model embeddings
 embeddings:
   model_name: "all-MiniLM-L6-v2"
-  # embedding_type: "sentence-transformer"
 
-# Cấu hình lưu trữ trong Chroma DB
+# Config ChromaDB
 chroma:
-  # persist_directory: "./chroma_db"
   collection_name: "knowledge"
   save_ids: true
   save_documents: false
   save_metadata: false
   client_settings:
-    # persist_directory: "./chroma_db"
     chroma_server_host: "localhost"
     chroma_server_http_port: 8000
 
-# Cấu hình tùy chọn tích hợp với PostgreSQL (nếu cần)
+# Config Postgresql
 postgresql:
   use_postgresql: true 
   host: "localhost"
@@ -35,24 +27,12 @@ postgresql:
   table_name: "vector_metadata"
   use_pgvector_extension: true  
 
-# Cấu hình LLM
-# llm:
-#   model_name: "meta-llama/Llama-3.2-3B"   # Tên model trên Hugging Face Hub
-#   provider: "huggingface"           
-#   temperature: 0.0
-#   max_tokens: 512
-#   huggingface:
-#     use_auth_token: true               
-
-
-
-# Tham số truy vấn
+   
 retrieval:
   top_k: 5
   similarity_threshold: 0.8 
 
-# Log
 logging:
-  level: "INFO"              # Mức độ log: DEBUG, INFO, WARNING, ERROR
+  level: "INFO"     
   log_dir: "./logs"
