+++ apps/chroma_db/docker-compose.yaml
@@ -18,8 +18,7 @@ services:
         - POSTGRES_DB=k1
         - POSTGRES_USER=k1
         - POSTGRES_PASSWORD=k1
-      volumes: 
-        - idx_knowledge_data:/var/lib/postgresql/data
+
 
   chroma:
     image: chromadb/chroma:latest
@@ -29,8 +28,4 @@ services:
     environment:
       - CHROMA_SERVER_HOST=0.0.0.0
       - CHROMA_SERVER_HTTP_PORT=8000
-    volumes:
-      - ./chroma_db:/chroma/.chroma
 
-volumes:
-  idx_knowledge_data:
