--- apps/chroma_db/ingestion/classifier_metadata.py
@@ -1,11 +1,0 @@
-import numpy as np
-from ingestion.embedder import build_embedder
-
-
-META_LABELS = {
-    "dsa": "..",
-    "mmt": "..",
-}
-
-
-
