+++ apps/chroma_db/test.py
@@ -1,7 +1,7 @@
 # api/main.py
 
 from fastapi import FastAPI
-from router.chroma_router import router
+from router.router import router
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI(
@@ -9,9 +9,8 @@ app = FastAPI(
     version="1.0.0"
 )
 
-# Mount router
 app.include_router(router)
-# bật CORS
+
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["*"],
@@ -20,7 +19,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# For manual run:  python api/main.py
+
 if __name__ == "__main__":
     import uvicorn
     uvicorn.run("test:app", host="0.0.0.0", port=8008, reload=True)
